Consolidate FileWrapper into archive_base and add additional file-like methods
