+++ nlds_processors/archiver/archive_base.py
@@ -11,6 +11,9 @@ __contact__ = 'neil.massey@stfc.ac.uk'
 from abc import ABC, abstractmethod
 import json
 from typing import Tuple, List, Dict
+from zlib import adler32
+
+from pyxrootd import client
 
 from nlds_processors.transferers.base_transfer import (BaseTransferConsumer, 
                                                        TransferError)
@@ -20,6 +23,47 @@ from nlds.details import PathDetails
 
 class ArchiveError(Exception):
     pass
+
+
+class AdlerisingXRDFile():
+    """Wrapper class around the XRootD.File object to make it act more like a 
+    regular python file object. This means it can interface with packages made 
+    for python, e.g. tarfile, BytesIO, minio. This also auto-calculates the 
+    adler32 checksum for all written/read bytes from the file, making 
+    implentation of checksums within the catalog feasible. 
+    """
+
+    def __init__(self, f: client.File, offset=0, length=0, checksum=1):
+        self.f = f
+        self.offset = offset
+        self.length = length
+        self.pointer = 0
+        self.checksum = checksum
+
+    def read(self, size):
+        """Read some number of bytes (size) from the file, offset by the current 
+        pointer position. Note this is wrapped by the adler checksumming but if 
+        used within a tarfile read this will not be done purely sequentially so 
+        will be essentially meaningless."""
+        status, result = self.f.read(offset=self.pointer, size=size)
+        if status.status != 0:
+            raise IOError(f"Unable to read from file f ({self.f})")
+        self.checksum = adler32(result, self.checksum)
+        self.pointer += size
+        return result
+    
+    def write(self, b):
+        # Update the checksum before we actually do the writing
+        self.checksum = adler32(b, self.checksum)
+        to_write = len(b)
+        self.f.write(b, offset=self.pointer, size=to_write)
+        return to_write
+    
+    def seek(self, whence: int) -> None:
+        self.pointer = whence
+    
+    def tell(self) -> int:
+        return self.pointer
 
 
 class BaseArchiveConsumer(BaseTransferConsumer, ABC):
